agents/pyproject.toml
[project]
name = "ai-factory-agents"
version = "0.1.0"
description = "AI Factory multi-agent system built on Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
